pyproject.toml
[tool.poetry]
name = "forex_data"
version = "0.1.0"
description = "Data aggregation for forex market data"
authors = ["Nicola Fiorato <fiorato.nicola@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "forex_data"}]

[tool.poetry.dependencies]
python = "~3.11"
spyder-kernels = ">=2.5.0"
pyaml = "^21.10.1"
pandas = ">=2.0.0"
matplotlib = ">=3.9.2"
requests = "^2.28.1"
mplfinance = "^0.12.9b7"
histdata = "^1.0"
polygon-api-client = "^1.5.0"
alpha-vantage = "^2.3.1"
mpl-finance = "^0.10.1"
setuptools = "^69.0.2"
scipy = "^1.11.4"
pyarrow = ">=14.0.2"
dotty-dict = {git = "https://github.com/nikfio/dotty_dict.git"}
polars = ">=0.20.7"
pytest = "^8.1.1"
loguru = "^0.7.2"
numpy  = "<2.0.0"
debugpy = "^1.8.5"
iteration-utilities = "^0.12.1"
bs4 = "^0.0.2"
flake8 = "^7.1.1"
mypy = "^1.14.1"
pandas-stubs = "^2.2.3.241126"
pytest-mypy = "^0.10.3"
pytest-flake8 = "^1.3.0"
sqlalchemy = "^2.0.38"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
